+++ modules/setLocalStorage.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 export const setLocalStorage = (array) => {
   localStorage.setItem('Books', JSON.stringify(array));
 };
